Use coop task instead of state machine and add error handling
We use the newly introduced coop task to make the handing of different
states easier. Additionally we handle errors that may occur after
changing the emissivity.

Changing the emissivity now has proper wait times (Between erase and
write).

If we get 5 errors in a row, we reset the MLX90614 by gooing to sleep,
waiting a bit and waking up again.